Remove old setup mode renderer and replace it with the react one
