+++ x-pack/plugins/monitoring/public/application/pages/apm/instances.tsx
@@ -15,7 +15,7 @@ import { useTable } from '../../hooks/use_table';
 import { ApmTemplate } from './apm_template';
 // @ts-ignore
 import { ApmServerInstances } from '../../../components/apm/instances';
-import { SetupModeRenderer } from '../../setup_mode/setup_mode_renderer';
+import { SetupModeRenderer } from '../../../components/renderers/setup_mode';
 import { SetupModeContext } from '../../../components/setup_mode/setup_mode_context';
 import { BreadcrumbContainer } from '../../hooks/use_breadcrumbs';
 import { APM_SYSTEM_ID } from '../../../../common/constants';
