+++ x-pack/plugins/monitoring/public/application/route_init.tsx
@@ -9,7 +9,7 @@ import { Route, Redirect, useLocation } from 'react-router-dom';
 import { useClusters } from './hooks/use_clusters';
 import { GlobalStateContext } from './contexts/global_state_context';
 import { getClusterFromClusters } from '../lib/get_cluster_from_clusters';
-import { isInSetupMode } from './setup_mode';
+import { isInSetupMode } from '../lib/setup_mode';
 import { LoadingPage } from './pages/loading_page';
 
 export interface ComponentProps {
