+++ x-pack/plugins/monitoring/public/components/renderers/setup_mode.d.ts
